requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
streamlit>=1.28.0
numpy>=1.24.0
pillow>=10.0.0
pytest>=7.4.0
watchdog>=3.0.0
